+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -14,6 +14,7 @@
 #include "EntryPointsNegCheck.h"
 #include "GuardCalledCheck.h"
 #include "GuardInForCheck.h"
+#include "RaddrConvBufLenCheck.h"
 #include "RaddrConvPureCheck.h"
 #include "ReserveLimitCheck.h"
 
@@ -34,6 +35,8 @@ public:
 	"hooks-guard-called");
     CheckFactories.registerCheck<GuardInForCheck>(
 	"hooks-guard-in-for");
+    CheckFactories.registerCheck<RaddrConvBufLenCheck>(
+        "hooks-raddr-conv-buf-len");
     CheckFactories.registerCheck<RaddrConvPureCheck>(
         "hooks-raddr-conv-pure");
     CheckFactories.registerCheck<ReserveLimitCheck>(