+++ clang-tools-extra/clang-tidy/hooks/RaddrConvBufLenCheck.h
@@ -1,0 +1,37 @@
+//===--- RaddrConvBufLenCheck.h - clang-tidy --------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_RADDRCONVBUFLENCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_RADDRCONVBUFLENCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+/// Function util_accid has upper limit on the length of its input
+/// (because it expects it to be a raddr) and fixed-size account ID
+/// output.
+class RaddrConvBufLenCheck : public ClangTidyCheck {
+public:
+  RaddrConvBufLenCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  const int64_t MAX_INPUT_SIZE = 49; // value copied from rippled source
+  const int64_t OUTPUT_SIZE = 20; // documented
+};
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_RADDRCONVBUFLENCHECK_H